requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.0
redis==5.2.0
celery[redis]==5.4.0
spacy==3.8.0
pytesseract==0.3.13
Pillow==11.0.0
httpx==0.27.0
beautifulsoup4==4.12.0
python-multipart==0.0.12
